PropertyHunterIA/requirements.txt
absl-py==0.8.0
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.1.1
future==0.18.1
gast==0.2.2
get==2019.4.13
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
joblib==0.13.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.17.2
oauthlib==3.1.0
opencv-contrib-python==3.4.3.18
openpyxl==3.0.0
opt-einsum==3.1.0
pandas==0.25.1
Pillow==6.2.0
post==2019.4.13
protobuf==3.9.2
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyEmail==0.0.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-telegram-bot==12.2.0
pytz==2019.2
PyYAML==5.1.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.21.3
scikit-plot==0.3.7
scipy==1.3.1
six==1.12.0
tagui==1.13.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
tornado==6.0.3
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
